🛠️ Active Task Indicator on Desktop Projects    - re-arranged task sort buttons in project details    - made task-item actions flat on projects page    - extracted an ActiveTask component    - extracted Tasks Actions mixins
